renamed BaseSet

+++ src/main/java/com/api/automation/model/Comment.java
@@ -12,9 +12,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Comment {
     @JsonProperty("postId")
-    private Integer postId;
+    private Long postId;
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
     @JsonProperty("name")
     private String name;
     @JsonProperty("email")
